+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -321,6 +321,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed various shader warning
 - Fixing unnecessary memory allocations in the ray tracing cluster build
 - Fixed duplicate column labels in LightEditor's light tab
+- Fixed white and dark flashes on scenes with very high or very low exposure when Automatic Exposure is being used.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled